chore: move to utils
