+++ packages/styled/src/util.ts
@@ -1,8 +1,10 @@
 import {
   applyEffect,
   customPropertyToken,
+  disabledSelector,
   filterObject,
   flatMapObject,
+  notDisabledSelector,
 } from '@charcoal-ui/utils'
 import { CharcoalAbstractTheme, EffectType, Key } from '@charcoal-ui/theme'
 import { CSSObject } from 'styled-components'
@@ -146,4 +148,17 @@ export function defineThemeVariables(
 
 export function isSupportedEffect(effect: Key): effect is EffectType {
   return ['hover', 'press', 'disabled'].includes(effect as string)
+}
+
+export const variable = (value: string) => `var(${value})`
+
+export function onEffectPseudo(effect: EffectType, css: CSSObject) {
+  return effect === 'hover'
+    ? { '&:hover': { [notDisabledSelector]: css } }
+    : effect === 'press'
+    ? { '&:active': { [notDisabledSelector]: css } }
+    : // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+    effect === 'disabled'
+    ? { [disabledSelector]: css }
+    : unreachable(effect)
 }
